+++ ricaun.RevitTest.Tests/TestsRevitTask.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 //[assembly: System.Reflection.AssemblyDescription("{\"TestAsync\":\"RevitTask\",\"TimeOut\":60.0}")]
 
 [assembly: System.Reflection.AssemblyMetadata("ricaun.RevitTest.Application.Tasks.Name", "RevitTask")]
-[assembly: System.Reflection.AssemblyMetadata("ricaun.RevitTest.Application.Tasks.Timeout", "5")]
+[assembly: System.Reflection.AssemblyMetadata("ricaun.RevitTest.Application.Tasks.Timeout", "0.05")]
 
 namespace ricaun.RevitTest.Tests
 {